[ruby] Update Ruby to 2.7.1. Fixes JB#50899.
Based on the fedora ruby package.
Bootstraping is now done by a recent ruby install on a host system
by 'generate_bootstrap.sh' and commited to the bootstrap directory.
This replicates the upstream bootstrap process for systems without
bootstrap but keeps the submodule.
The bootstraping can be disabled by a rpm build condition.

Signed-off-by: Björn Bidar <bjorn.bidar@jolla.com>