[ruby] Fix compilation with newer bison. JB#57049
The latest update with commit d3b9dced7 regenerated the parse.h
with bison 3.5.1 when the earlier was with 3.6.4. Some enum
values were missed which made the compilation now fail with Bison 3.8.

Would be good to regenerate the bootstrap but as minimum effort
fix now just manually adding them back. Otherwise the enum values
and their numbers are the same.